+++ tomcat/src/main/java/org/apache/catalina/controller/RegisterController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package org.apache.catalina.controller;
 
 import java.io.IOException;
 import java.util.Map;
