+++ tomcat/src/main/java/org/apache/catalina/controller/StaticResourceController.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.controller;
+package org.apache.catalina.controller;
 
 import java.io.IOException;
 import org.apache.coyote.http11.exception.ResourceLoadingException;
