+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -11,8 +11,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import nextstep.jwp.exception.UncheckedServletException;
+import org.apache.catalina.ControllerMapper;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.controller.ControllerMapper;
 import org.apache.coyote.http11.header.HeaderName;
 import org.apache.coyote.http11.header.Headers;
 import org.apache.coyote.http11.request.HttpRequest;
